Update pipeline complete untested yet

+++ src/DataPipeLine/Scripts/singleday_pdf_webscrapper.py
@@ -1,10 +1,12 @@
 from urllib.request import urlopen, urlretrieve
 from urllib.parse import urljoin, quote
-from datetime import date
+import datetime
+import os
 from bs4 import BeautifulSoup
 
 incomplete_url = "http://www.epid.gov.lk/web/images/pdf/corona_virus_report/sitrep-sl-en-"
-[day, month, year] = str(date.today())[2:].split('-')[::-1]
+today = datetime.datetime.now()
+current_date = [today.month,today.day]
 
 Directory='F:/Deeplearning/Datathon-1.0'
 
@@ -30,6 +32,14 @@ for link in soup.select('p a'):
     parts = filename.split('-')
     dw_name = parts[4][:2] + '-' + parts[3] + '.pdf'
 
+    month = parts[4][:2]
+    day = parts[3]
+
+    print(current_date)
+    print([int(month),int(day)])
+    print([int(month),int(day)] == current_date)
+    print(not os.path.exists(Directory + '/data/daily_pdf/' + dw_name))
+    print('*** \n \n')
     #if (parts[1] == 'sl' and parts[2] == 'en' and filename != 'dailyreportcoronavirussin.pdf'):
     #    href = urljoin(url, quote(href))
     #    filepath = urlretrieve(href, './Datathon-1.0/data/daily_pdf/' + dw_name)
@@ -42,13 +52,15 @@ for link in soup.select('p a'):
             # print(href)
             #print('../../../data/daily_pdf/'+dw_name)
             #filepath = urlretrieve(href,'./Datathon-1.0/data/daily_pdf/' + dw_name)
-            try:
-                filepath = urlretrieve(href, Directory + '/data/daily_pdf/' + dw_name)
-                print('Sucess :- ', filepath[0], ' created')
-            except:
-                print('url:- ', href)
-                print('date:- ', parts[4][:2] + '-' + parts[3])
-                print('failed to download')
+            if([int(month),int(day)] == current_date and (not os.path.exists(Directory + '/data/daily_pdf/' + dw_name))):
+                try:
+                    filepath = urlretrieve(href, Directory + '/data/daily_pdf/' + dw_name)
+                    print('Sucess :- ', filepath[0], ' created')
+                except:
+                    print('url:- ', href)
+                    print('date:- ', parts[4][:2] + '-' + parts[3])
+                    print('failed to download')
+     
     except:
         print(filename)
